Modify Newman test for login
