+++ tests/tests/backend_collection.json
@@ -45,7 +45,7 @@
 				],
 				"body": {
 					"mode": "raw",
-					"raw": "{\n    \"username\": \"admin\",\n    \"password\": \"admin\"\n}"
+					"raw": "{\n    \"username\": \"admin\",\n    \"password\": \"password\"\n}"
 				},
 				"url": {
 					"raw": "{{baseUrl}}/login",
